refactor: Move the useThemeLabel hook to Hooks folder
